+++ src/components/Home.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 function Home() {
   return (
     <div className="flex flex-col justify-center items-center h-screen gap-6">
-      <Link to={"/login"}>
+      <Link to={"/user/user-profile"}>
         <p>user Page</p>
       </Link>
       <Link to={"/posts"}>
