+++ src/components/index.js
@@ -1,0 +1,6 @@
+import Home from "./Home";
+import Navbar from "./Navbar";
+import Footer from "./Footer";
+import NotFound from "./NotFound";
+
+export { Home, Navbar, Footer, NotFound };
