+++ src/constants/actionTypes.js
@@ -1,0 +1,44 @@
+// users action types
+export const LOGOUT_USER = "LOGOUT_USER";
+
+export const SIGNUP_USER_REQUEST = "SIGNUP_USER_REQUEST";
+export const SIGNUP_USER_SUCCESS = "SIGNUP_USER_SUCCESS";
+export const SIGNUP_USER_FAILURE = "SIGNUP_USER_FAILURE";
+
+export const LOGIN_USER_REQUEST = "LOGIN_USER_REQUEST";
+export const LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS";
+export const LOGIN_USER_FAILURE = "LOGIN_USER_FAILURE";
+
+export const FETCH_USERS_REQUEST = "FETCH_USERS_REQUEST";
+export const FETCH_USERS_SUCCESS = "FETCH_USERS_SUCCESS";
+export const FETCH_USERS_FAILURE = "FETCH_USERS_FAILURE";
+
+// posts action types
+
+export const FETCH_POSTS_REQUEST = "FETCH_POSTS_REQUEST";
+export const FETCH_POSTS_SUCCESS = "FETCH_POSTS_SUCCESS";
+export const FETCH_POSTS_FAILURE = "FETCH_POSTS_FAILURE";
+
+export const ADD_POST_REQUEST = "ADD_POST_REQUEST";
+export const ADD_POST_SUCCESS = "ADD_POST_SUCCESS";
+export const ADD_POST_FAILURE = "ADD_POST_FAILURE";
+
+export const DELETE_POST_REQUEST = "DELETE_POST_REQUEST";
+export const DELETE_POST_SUCCESS = "DELETE_POST_SUCCESS";
+export const DELETE_POST_FAILURE = "DELETE_POST_FAILURE";
+
+export const FETCH_POST_REQUEST = "FETCH_POST_REQUEST";
+export const FETCH_POST_SUCCESS = "FETCH_POST_SUCCESS";
+export const FETCH_POST_FAILURE = "FETCH_POST_FAILURE";
+
+// Todos action types
+
+export const GET_TODOS = "GET_TODOS";
+export const GET_TODO = "GET_TODO";
+export const ADD_TODO = "ADD_TODO";
+export const UPDATE_TODO = "ADD_TODO";
+export const DELETE_TOD = "DELETE_TODO";
+
+// dark mode
+
+export const TOGGLE_DARK_MODE = "TOGGLE_DARK_MODE";
