+++ src/features/auth/LoginPage.jsx
@@ -1,0 +1,3 @@
+import { LoginForm as LoginPage } from "./components";
+
+export default LoginPage;
