+++ src/features/auth/RegistrationPage.jsx
@@ -1,0 +1,3 @@
+import { RegistrationForm as RegistrationPage } from "./components";
+
+export default RegistrationPage;
