+++ src/features/auth/components/Logout.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Logout() {
+  return <div>Logout</div>;
+}
+
+export default Logout;
