+++ src/features/auth/components/index.js
@@ -1,0 +1,6 @@
+import Auth from "./Auth";
+import LoginForm from "./LoginForm";
+import Logout from "./Logout";
+import RegistrationForm from "./RegistrationForm";
+
+export { LoginForm, RegistrationForm, Logout, Auth };
