+++ src/features/auth/index.js
@@ -1,0 +1,5 @@
+import LoginPage from "./LoginPage";
+import RegistrationPage from "./RegistrationPage";
+import authReducer from "./authReducer";
+
+export { LoginPage, RegistrationPage, authReducer };
