--- src/features/posts/postActionTypes.js
@@ -1,15 +1,0 @@
-export const FETCH_POSTS_REQUEST = "FETCH_POSTS_REQUEST";
-export const FETCH_POSTS_SUCCESS = "FETCH_POSTS_SUCCESS";
-export const FETCH_POSTS_FAILURE = "FETCH_POSTS_FAILURE";
-
-export const ADD_POST_REQUEST = "ADD_POST_REQUEST";
-export const ADD_POST_SUCCESS = "ADD_POST_SUCCESS";
-export const ADD_POST_FAILURE = "ADD_POST_FAILURE";
-
-export const DELETE_POST_REQUEST = "DELETE_POST_REQUEST";
-export const DELETE_POST_SUCCESS = "DELETE_POST_SUCCESS";
-export const DELETE_POST_FAILURE = "DELETE_POST_FAILURE";
-
-export const FETCH_POST_REQUEST = "FETCH_POST_REQUEST";
-export const FETCH_POST_SUCCESS = "FETCH_POST_SUCCESS";
-export const FETCH_POST_FAILURE = "FETCH_POST_FAILURE";
