+++ src/features/user/components/index.js
@@ -1,0 +1,3 @@
+import UserProfile from "./UserProfile";
+
+export { UserProfile };
