+++ src/features/user/index.js
@@ -1,0 +1,3 @@
+import UserProfilePage from "./UserProfilePage";
+
+export { UserProfilePage };
