+++ src/pages/Home.jsx
@@ -1,7 +1,2 @@
-import React from "react";
-
-function Home() {
-  return <div>Home</div>;
-}
-
+import { Home } from "../components";
 export default Home;
