+++ src/pages/Loginpage.jsx
@@ -1,0 +1,3 @@
+import { LoginPage } from "../features/auth";
+
+export default LoginPage;
