+++ src/pages/PostsDetailsPage.jsx
@@ -1,3 +1,3 @@
-import { PostsDetailsPage } from "../features/posts";
+import { PostsDetails as PostsDetailsPage } from "../features/posts";
 
 export default PostsDetailsPage;
