+++ src/pages/Registrationpage.jsx
@@ -1,0 +1,3 @@
+import { RegistrationPage } from "../features/auth";
+
+export default RegistrationPage;
