+++ src/pages/UserProfilePage.jsx
@@ -1,0 +1,3 @@
+import { UserProfilePage } from "../features/user";
+
+export default UserProfilePage;
