+++ src/store/rootReducer.js
@@ -1,8 +1,8 @@
 import { combineReducers } from "redux";
-import todoReducer from "./todoReducer";
-import authReducer from "./authReducer";
-import { postReducer } from "./features/posts";
-import darkModeReducer from "./darkModeReducer";
+import todoReducer from "../todoReducer";
+import { authReducer } from "../features/auth";
+import { postReducer } from "../features/posts";
+import darkModeReducer from "../darkModeReducer";
 
 const rootReducer = combineReducers({
   todoState: todoReducer,
